fix(guard): remove workflow capability import cycles
